requirements.txt
google-cloud-bigquery>=3.25
google-auth>=2.35
requests>=2.32
